+++ addons/ammunition/CfgAmmo.hpp
@@ -12,7 +12,6 @@ class CfgAmmo {
     class B_12Gauge_Pellets_Submunition_Deploy;
     class B_9x21_Ball_Tracer_Yellow; // 9x19
     class B_545x39_Ball_F; // 5.45x39
-    class CUP_B_570x28_Ball_Tracer_Yellow; // 5.7x28
     class B_45ACP_Ball; // .45ACP
     class B_580x42_Ball_F; // 5.8x42
     class B_556x45_Ball; // 5.56x45
@@ -20,8 +19,6 @@ class CfgAmmo {
     class B_762x39_Ball_F; // 7.62x39
     class B_762x51_Ball; // 7.62x51
     class B_762x54_Ball; // 7.62x54R
-    class HLC_300Blackout_Ball; // .300 Blackout Ball (Formerly subsonic)
-    class CUP_B_9x39_SP5; // 9x39
 
     // 12G
     class CLASS(12g_Pellets_Submunition): B_12Gauge_Pellets_Submunition {
@@ -122,19 +119,6 @@ class CfgAmmo {
     // 10mm
     class CLASS(10mm_Ball): CLASS(357_Magnum) {
         hit = 12;
-    };
-
-    // 5.7x28mm
-    class CLASS(57x28_Ball): CUP_B_570x28_Ball_Tracer_Yellow {
-        aiAmmoUsageFlags = "64 + 128";
-        caliber = 1.5;
-        hit = 12;
-        MACRO_TRACERS
-        ACE_caliber = 5.7;
-        ACE_bulletLength = 21.6;
-        ACE_bulletMass = 2;
-        ACE_muzzleVelocities[] = {716, 776};
-        ACE_barrelLengths[] = {264, 407};
     };
 
     // 4.6x30mm (Comparable to 5.7x28mm but slightly worse performing penetration.)
@@ -285,43 +269,5 @@ class CfgAmmo {
     class CLASS(762x54r_AP): CLASS(762x54r_EPR) {
         caliber = 2.5;
         hit = 14.5;
-    };
-
-
-    // .300AAC - Not subsonic anymore - Ridiculous drop at 400m
-    class CLASS(300AAC_Ball): HLC_300Blackout_Ball {
-        aiAmmoUsageFlags = "64 + 128";
-        caliber = 0.19;
-        hit = 12;
-        MACRO_TRACERS
-    };
-    class CLASS(300AAC_EPR): CLASS(300AAC_Ball) {
-        caliber = 1.2;
-        hit = 13;
-    };
-    class CLASS(300AAC_AP): CLASS(300AAC_EPR) {
-        caliber = 1.4;
-        hit = 13.5;
-    };
-
-    // 9x39mm - Subsonic
-    class CLASS(9x39_Ball): CUP_B_9x39_SP5 {
-        aiAmmoUsageFlags = "64 + 128";
-        caliber = 0.15;
-        hit = 10;
-        MACRO_TRACERS
-    };
-    class CLASS(9x39_EPR): CLASS(9x39_Ball) {
-        caliber = 1.1;
-        hit = 11;
-    };
-    class CLASS(9x39_AP): CLASS(9x39_EPR) {
-        caliber = 1.3;
-        hit = 11.5;
-    };
-
-    // MELB Tracer Size
-    class B_762x51_Ball_MELB: B_762x51_Ball {
-        tracerScale = 1;
     };
 };
