+++ addons/ammunition/CfgMagazines.hpp
@@ -2,95 +2,43 @@ class CfgMagazines {
     // Baseclasses
     class CA_Magazine;
     class 2Rnd_12Gauge_Pellets;
-    class CUP_5Rnd_B_Saiga12_Buck_00;
     class 6Rnd_12Gauge_Pellets;
     class 6Rnd_12Gauge_Slug;
     class 30Rnd_9x21_Mag;
-    class CUP_30Rnd_9x19_MP5;
-    class CUP_30Rnd_9x19_Vityaz;
-    class CUP_64Rnd_9x19_Bizon_M;
-    class CUP_40Rnd_46x30_MP7;
     class 30Rnd_545x39_Mag_F;
-    class CUP_30Rnd_545x39_Fort224_M;
-    class CUP_60Rnd_545x39_AK74M_M;
     class 30Rnd_45ACP_Mag_SMG_01;
     class 50Rnd_570x28_SMG_03;
     class 30Rnd_580x42_Mag_F;
     class 100Rnd_580x42_Mag_F;
     class 30Rnd_556x45_Stanag;
-    class CUP_30Rnd_556x45_AK;
-    class CUP_25Rnd_556x45_Famas;
-    class hlc_30Rnd_556x45_EPR_sg550;
-    class hlc_30Rnd_556x45_B_AUG;
-    class hlc_40Rnd_556x45_B_AUG;
-    class CUP_30Rnd_TE1_Green_Tracer_556x45_G36;
-    class CUP_60Rnd_556x45_SureFire;
     class 150Rnd_556x45_Drum_Mag_F;
-    class CUP_100Rnd_556x45_BetaCMag;
     class 200Rnd_556x45_Box_F;
-    class CUP_100Rnd_TE4_Green_Tracer_556x45_M249;
-    class CUP_200Rnd_TE4_Green_Tracer_556x45_M249_Pouch;
-    class CUP_200Rnd_TE4_Green_Tracer_556x45_M249;
     class 30Rnd_65x39_caseless_mag; // MX
     class 100Rnd_65x39_caseless_mag;
     class 30Rnd_65x39_caseless_green; // Katiba - Type-115
     class 30Rnd_65x39_caseless_msbs_mag; // MSBS
     class 200Rnd_65x39_cased_Box;
     class 20Rnd_650x39_Cased_Mag_F;
-    class CUP_30Rnd_680x43_Stanag;
     class 30Rnd_762x39_AK12_Mag_F; // Plastic AK Mag
     class 30Rnd_762x39_Mag_F; // Metal AK Mag
     class 75rnd_762x39_AK12_Mag_F;
-    class CUP_30Rnd_762x39_CZ807;
     class 20Rnd_762x51_Mag;
-    class CUP_20Rnd_762x51_FNFAL_M;
-    class CUP_20Rnd_762x51_B_SCAR_bkl;
-    class hlc_100Rnd_762x51_B_M60E4;
     class 150Rnd_762x54_Box;
     class 10Rnd_762x54_Mag;
-    class hlc_29rnd_300BLK_STANAG;
-    class CUP_30Rnd_9x39_SP5_VIKHR_M;
-    class 35Rnd_556x45_Velko_reload_tracer_yellow_lxWS;
-    class 8Rnd_12Gauge_AA40_Pellets_lxWS;
-    class 20Rnd_12Gauge_AA40_Pellets_lxWS;
-    class 20Rnd_762x51_slr_lxWS;
-    class 100Rnd_762x51_S77_Yellow_lxWS;
 
     // Handgun baseclasses
     class 16Rnd_9x21_Mag;
-    class CUP_13Rnd_9x19_Browning_HP;
-    class CUP_18Rnd_9x19_Phantom;
-    class CUP_17Rnd_9x19_glock17;
-    class hlc_15Rnd_9x19_B_P226;
-    class CUP_17Rnd_9x19_M17_Black;
-    class CUP_15Rnd_9x19_M9;
-    class hlc_10Rnd_9x19_B_P239;
-    class CUP_7Rnd_45ACP_1911;
     class 11Rnd_45ACP_Mag;
-    class CUP_12Rnd_45ACP_mk23;
-    class hlc_12Rnd_357SIG_B_P226;
-    class hlc_8Rnd_357SIG_B_P239;
 
-    // Rifles
-    #include "magazines\12g.hpp"
-    #include "magazines\9x19.hpp"
-    #include "magazines\9x39.hpp"
-    #include "magazines\45ACP.hpp"
-    #include "magazines\46x30.hpp"
-    #include "magazines\57x28.hpp"
-    #include "magazines\58x42.hpp"
-    #include "magazines\65x39.hpp"
-    #include "magazines\68x43.hpp"
-    #include "magazines\300Blackout.hpp"
-    #include "magazines\545x39.hpp"
-    #include "magazines\556x45.hpp"
-    #include "magazines\762x39.hpp"
-    #include "magazines\762x51.hpp"
-    #include "magazines\762x54r.hpp"
-
-    // Handguns
-    #include "handgun_magazines\9x19.hpp"
-    #include "handgun_magazines\45ACP.hpp"
-    #include "handgun_magazines\357.hpp"
-    #include "handgun_magazines\10mm.hpp"
+    #include "bi\magazines\12g.hpp"
+    #include "bi\magazines\45ACP.hpp"
+    #include "bi\magazines\545x39.hpp"
+    #include "bi\magazines\556x45.hpp"
+    #include "bi\magazines\57x28.hpp"
+    #include "bi\magazines\58x42.hpp"
+    #include "bi\magazines\65x39.hpp"
+    #include "bi\magazines\762x39.hpp"
+    #include "bi\magazines\762x51.hpp"
+    #include "bi\magazines\762x54r.hpp"
+    #include "bi\magazines\9x19.hpp"
 };
