+++ addons/ammunition/bi/magazines/58x42.hpp
