+++ addons/ammunition/bi/magazines/65x39.hpp
