+++ addons/ammunition/bi/magazines/762x54r.hpp
