+++ addons/ammunition/cup/magazines/68x43.hpp
