+++ addons/ammunition/cup/magazines/9x39.hpp
