+++ addons/ammunition/melb/script_component.hpp
@@ -1,0 +1,6 @@
+#define COMPONENT ammunition_melb
+#define COMPONENT_BEAUTIFIED Ammunition MELB
+#include "\x\tacgt\addons\main\script_mod.hpp"
+#include "\x\tacgt\addons\main\script_macros.hpp"
+
+#include "..\script_macros.hpp"
