+++ addons/ammunition/niarms/CfgMagazines.hpp
@@ -1,0 +1,18 @@
+class CfgMagazines {
+    class hlc_30Rnd_556x45_EPR_sg550;
+    class hlc_30Rnd_556x45_B_AUG;
+    class hlc_40Rnd_556x45_B_AUG;
+    class hlc_100Rnd_762x51_B_M60E4;
+    class hlc_29rnd_300BLK_STANAG;
+    class hlc_15Rnd_9x19_B_P226;
+    class hlc_10Rnd_9x19_B_P239;
+    class hlc_12Rnd_357SIG_B_P226;
+    class hlc_8Rnd_357SIG_B_P239;
+
+    #include "magazines\10mm.hpp"
+    #include "magazines\300Blackout.hpp"
+    #include "magazines\357.hpp"
+    #include "magazines\556x45.hpp"
+    #include "magazines\762x51.hpp"
+    #include "magazines\9x19.hpp"
+};
