+++ addons/ammunition/niarms/magazines/10mm.hpp
