+++ addons/ammunition/niarms/magazines/300Blackout.hpp
