+++ addons/ammunition/niarms/magazines/357.hpp
