+++ addons/ammunition/niarms/magazines/9x19.hpp
@@ -1,0 +1,19 @@
+// 9x19 SIG
+class hlc_15Rnd_9x19_B_P226;
+class CLASS(15Rnd_9x19_P226_Ball): hlc_15Rnd_9x19_B_P226 {
+    MACRO_SCOPE
+    ammo = QCLASS(9x19_Ball);
+    descriptionShort = "9x19mm Ball Reload Tracer";
+    displayName = "9mm 15Rnd P226 (Ball)";
+    lastRoundsTracer = 2;
+};
+
+// 9x19 SIG P239
+class hlc_10Rnd_9x19_B_P239;
+class CLASS(10Rnd_9x19_P239_Ball): hlc_10Rnd_9x19_B_P239 {
+    MACRO_SCOPE
+    ammo = QCLASS(9x19_Ball);
+    descriptionShort = "9x19mm Ball Reload Tracer";
+    displayName = "9mm 10Rnd P239 (Ball)";
+    lastRoundsTracer = 2;
+};
