+++ addons/ammunition/niarms/script_component.hpp
@@ -1,0 +1,6 @@
+#define COMPONENT ammunition_hlc
+#define COMPONENT_BEAUTIFIED Ammunition NIArms
+#include "\x\tacgt\addons\main\script_mod.hpp"
+#include "\x\tacgt\addons\main\script_macros.hpp"
+
+#include "..\script_macros.hpp"
