+++ addons/ammunition/pi/script_component.hpp
@@ -1,0 +1,6 @@
+#define COMPONENT ammunition_pi
+#define COMPONENT_BEAUTIFIED Ammunition PI
+#include "\x\tacgt\addons\main\script_mod.hpp"
+#include "\x\tacgt\addons\main\script_macros.hpp"
+
+#include "..\script_macros.hpp"
