+++ addons/ammunition/ws/CfgMagazines.hpp
@@ -1,0 +1,11 @@
+class CfgMagazines {
+    class 35Rnd_556x45_Velko_reload_tracer_yellow_lxWS;
+    class 8Rnd_12Gauge_AA40_Pellets_lxWS;
+    class 20Rnd_12Gauge_AA40_Pellets_lxWS;
+    class 20Rnd_762x51_slr_lxWS;
+    class 100Rnd_762x51_S77_Yellow_lxWS;
+
+    #include "magazines\12g.hpp"
+    #include "magazines\556x45.hpp"
+    #include "magazines\762x51.hpp"
+};
